Adjust .gitignore to ignore game assets/artifacts not in the repository
This adds the files `WT Silver.exe`, `WT Silver.py`, "lib/" and "renpy"
to the .gitignore file. These are all not included in the repository,
and hence can be savely ignored, but are needed to execute the game.
And already ignored asset is the `game/music` directory, previously
listed under the eclipse related ignore list, now listed and regrouped
at the top of the .gitignore file.